pyproject.toml
[tool.poetry]
name = "nix-survey-analysis"
version = "0.1.0"
description = ""
authors = ["Guillaume Desforges <guillaume.desforges.pro@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "nixos"}]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.25.2"
polars = "^0.18.15"
matplotlib = "^3.7.2"
pydantic = "^2.2.1"
pyyaml = "^6.0.1"
pyarrow = "^13.0.0"
pandas = "^2.0.3"
tabulate = "^0.9.0"


[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
ruff = "^0.0.285"
pyright = "^1.1.324"
ipykernel = "^6.25.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
select = ["E", "F", "I"]
ignore = ["E501"]
